Split into nested tests, rename, add todos
